Add gradle configurations for jsonpath operation
