+++ ballerina/build.gradle
@@ -69,6 +69,9 @@ task updateTomlFiles {
     doLast {
         def newConfig = ballerinaTomlFilePlaceHolder.text.replace("@project.version@", project.version)
         newConfig = newConfig.replace("@toml.version@", tomlVersion)
+        newConfig = newConfig.replace("@jsonpath.version@", project.javaJsonPathVersion)
+        newConfig = newConfig.replace("@jsonsmart.version@", project.javaJsonSmartVersion)
+        newConfig = newConfig.replace("@accessors.version@", project.javaAccessorsSmartVersion)
         ballerinaTomlFile.text = newConfig
 
         def newCompilerPluginToml = compilerPluginTomlFilePlaceHolder.text.replace("@project.version@", project.version)
