+++ ballerina/json_api.bal
@@ -66,10 +66,6 @@ public type Options record {|
     boolean allowDataProjection = true;
 |};
 
-# Represents the error type of the ballerina/data.jsondata module. This error type represents any error that can occur
-# during the execution of jsondata APIs.
-public type Error distinct error;
-
 # Defines the name of the JSON Object key.
 #
 # + value - The name of the JSON Object key
