+++ gradle.properties
@@ -12,4 +12,7 @@ githubJohnrengelmanShadowVersion=8.1.1
 underCouchDownloadVersion=4.0.4
 researchgateReleaseVersion=2.8.0
 ballerinaGradlePluginVersion=2.0.1
+javaJsonPathVersion=2.7.0
+javaJsonSmartVersion=2.4.7
+javaAccessorsSmartVersion=2.4.7
 stdlibIoVersion=1.6.0
