+++ native/build.gradle
@@ -33,6 +33,7 @@ dependencies {
     implementation group: 'org.ballerinalang', name: 'ballerina-lang', version: "${ballerinaLangVersion}"
     implementation group: 'org.ballerinalang', name: 'ballerina-runtime', version: "${ballerinaLangVersion}"
     implementation group: 'org.ballerinalang', name: 'value', version: "${ballerinaLangVersion}"
+    implementation group: 'com.jayway.jsonpath', name: 'json-path', version: "${javaJsonPathVersion}"
 }
 
 checkstyle {
