+++ native/src/main/java/io/ballerina/lib/data/jsonpath/BJsonPath.java
@@ -16,7 +16,7 @@
  *  under the License.
  */
 
-package io.ballerina.stdlib.jsonpath;
+package io.ballerina.lib.data.jsonpath;
 
 import com.jayway.jsonpath.Configuration;
 import com.jayway.jsonpath.JsonPathException;
@@ -26,8 +26,7 @@ import io.ballerina.runtime.api.values.BError;
 import io.ballerina.runtime.api.values.BObject;
 
 import static com.jayway.jsonpath.JsonPath.using;
-
-import static io.ballerina.stdlib.jsonpath.Utils.convertRawTemplateToString;
+import static io.ballerina.lib.data.jsonpath.Utils.convertRawTemplateToString;
 
 /**
  * Provides function implementations of json-path.
@@ -37,7 +36,6 @@ import static io.ballerina.stdlib.jsonpath.Utils.convertRawTemplateToString;
 public class BJsonPath {
     public static final Configuration BJSON_CONFIGURATION = Configuration
             .builder()
-            .mappingProvider(new BJsonMappingProvider())
             .jsonProvider(new BJsonProvider())
             .build();
